refactor: move simplify list functionality into companion object
