+++ exact-numbers/src/test/kotlin/common/UtilsTest.kt
@@ -1,4 +1,4 @@
-package shared
+package common
 
 import assertDivByZero
 import java.math.BigDecimal
