+++ exact-numbers/src/test/kotlin/exactnumbers/irrationals/pi/simplifyListTests.kt
@@ -1,0 +1,56 @@
+package exactnumbers.irrationals.pi
+
+import exactnumbers.irrationals.log.Log
+import kotlin.test.assertEquals
+import kotlin.test.assertFails
+
+internal fun runSimplifyListTests() {
+    // error
+    assertFails { Pi.simplifyList(listOf(Pi(), Log.ONE)) }
+
+    // equal
+    var expected: List<Pi> = listOf()
+
+    var pis: List<Pi> = listOf()
+    assertEquals(expected, Pi.simplifyList(pis))
+
+    pis = listOf(Pi(), Pi(true))
+    assertEquals(expected, Pi.simplifyList(pis))
+
+    pis = listOf(Pi(true), Pi(), Pi(true), Pi(), Pi(), Pi(true))
+    assertEquals(expected, Pi.simplifyList(pis))
+
+    // positive
+    pis = listOf(Pi())
+    expected = listOf(Pi())
+    assertEquals(expected, Pi.simplifyList(pis))
+
+    pis = listOf(Pi(), Pi(), Pi())
+    expected = listOf(Pi(), Pi(), Pi())
+    assertEquals(expected, Pi.simplifyList(pis))
+
+    pis = listOf(Pi(), Pi(), Pi(true))
+    expected = listOf(Pi())
+    assertEquals(expected, Pi.simplifyList(pis))
+
+    pis = listOf(Pi(), Pi(), Pi(), Pi(true), Pi(), Pi(true))
+    expected = listOf(Pi(), Pi())
+    assertEquals(expected, Pi.simplifyList(pis))
+
+    // negative
+    pis = listOf(Pi(true))
+    expected = listOf(Pi(true))
+    assertEquals(expected, Pi.simplifyList(pis))
+
+    pis = listOf(Pi(true), Pi(true), Pi(true))
+    expected = listOf(Pi(true), Pi(true), Pi(true))
+    assertEquals(expected, Pi.simplifyList(pis))
+
+    pis = listOf(Pi(true), Pi(true), Pi())
+    expected = listOf(Pi(true))
+    assertEquals(expected, Pi.simplifyList(pis))
+
+    pis = listOf(Pi(true), Pi(true), Pi(true), Pi(), Pi(true), Pi())
+    expected = listOf(Pi(true), Pi(true))
+    assertEquals(expected, Pi.simplifyList(pis))
+}
